pyproject.toml
[project]
name = "midi2image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "certifi>=2025.6.15",
    "ctkmessagebox>=2.7",
    "customtkinter>=5.2.2",
    "mido>=1.3.3",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "tkinterdnd2>=0.4.3",
]
